+++ logprep/registry.py
@@ -9,6 +9,7 @@ from logprep.connector.console.output import ConsoleOutput
 from logprep.connector.dummy.input import DummyInput
 from logprep.connector.dummy.output import DummyOutput
 from logprep.connector.elasticsearch.output import ElasticsearchOutput
+from logprep.connector.http.input import HttpConnector
 from logprep.connector.json.input import JsonInput
 from logprep.connector.jsonl.input import JsonlInput
 from logprep.connector.jsonl.output import JsonlOutput
@@ -73,6 +74,7 @@ class Registry:
         "elasticsearch_output": ElasticsearchOutput,
         "jsonl_output": JsonlOutput,
         "opensearch_output": OpensearchOutput,
+        "http_input": HttpConnector,
     }
 
     @classmethod